+++ src/window/custom.rs
@@ -16,7 +16,8 @@ impl CustomWindowFactory {
             .with_inner_size(LogicalSize::new(options.width, options.height))
             .with_resizable(options.resizable)
             .with_decorations(false)
-            .with_transparent(true);
+            .with_transparent(true)
+            .with_window_icon(options.icon.clone());
 
         if let Some((x, y)) = options.position {
             builder = builder.with_position(PhysicalPosition::new(x, y));